feat: member info edit and exception handling impl.

+++ server/src/main/java/pre045/board_service/member/service/MemberService.java
@@ -5,12 +5,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pre045.board_service.exception.BusinessLogicException;
-import pre045.board_service.exception.ExceptionCode;
 import pre045.board_service.member.dto.MemberLoginDto;
+import pre045.board_service.member.dto.MemberPatchDto;
 import pre045.board_service.member.dto.MemberPostDto;
 import pre045.board_service.member.dto.MemberResponseDto;
 import pre045.board_service.member.entity.Member;
@@ -58,30 +59,34 @@ public class MemberService {
         return MemberResponseDto.of(memberRepository.save(member));
     }
 
+    //로그인
     public TokenDto login(MemberLoginDto loginDto) {
         //email, pw -> Authentication Token
-        UsernamePasswordAuthenticationToken authenticationToken = loginDto.toAuthentication();
+        UsernamePasswordAuthenticationToken authenticationToken = toAuthentication(loginDto.getEmail(), loginDto.getPassword());
 
+        Authentication authentication;
 
         //email, pw 검증
-        Authentication authentication = authBuilder.getObject().authenticate(authenticationToken);
-
+        try {
+            authentication = authBuilder.getObject().authenticate(authenticationToken);
+        } catch (AuthenticationException e) {
+            throw new BusinessLogicException(MEMBER_NOT_EXIST);
+        }
 
         //토큰 생성
         TokenDto token = tokenProvider.createToken(authentication);
-
 
         RefreshToken refreshToken = RefreshToken.builder()
                 .tokenKey(authentication.getName()) // tokenKey == memberId 로 설정
                 .tokenValue(token.getRefreshToken())
                 .build();
 
-
         refreshTokenRepository.save(refreshToken);
 
         return token;
     }
 
+    //액세스 토큰 재발급
     public TokenDto reissue(TokenRequestDto tokenRequestDto) {
         //Refresh Token 검증
         if (!tokenProvider.validateToken(tokenRequestDto.getRefreshToken())) {
@@ -112,24 +117,53 @@ public class MemberService {
     }
 
 
+    //로그아웃
     public void logout() {
         Long memberId = SecurityUtil.getCurrentMemberId();
         RefreshToken refreshToken = verifyRefreshToken(memberId.toString());
         refreshTokenRepository.delete(refreshToken);
     }
 
-    public void deleteMember() {
-        Long memberId = SecurityUtil.getCurrentMemberId();
+    //회원 탈퇴
+    public void deleteMember(Long memberId) {
+        Long authMemberId = SecurityUtil.getCurrentMemberId();
+
+        if (!memberId.equals(authMemberId)) {
+            throw new BusinessLogicException(NOT_MATCH_TOKEN_WITH_MEMBER);
+        }
+
 
         //회원 정보 삭제
-        Member foundMember = findVerifiedMember(memberId);
+        Member foundMember = findVerifiedMember(authMemberId);
         memberRepository.delete(foundMember);
 
         //refresh token 삭제
-        RefreshToken refreshToken = verifyRefreshToken(memberId.toString());
+        RefreshToken refreshToken = verifyRefreshToken(authMemberId.toString());
         refreshTokenRepository.delete(refreshToken);
     }
 
+    //회원 정보 수정
+    public MemberResponseDto editInfo(MemberPatchDto patchDto) {
+        Long memberId = SecurityUtil.getCurrentMemberId();
+        Member foundMember = findVerifiedMember(memberId);
+
+        UsernamePasswordAuthenticationToken authenticationToken = toAuthentication(foundMember.getEmail(), patchDto.getPrePassword());
+
+        try {
+            authBuilder.getObject().authenticate(authenticationToken);
+        } catch (AuthenticationException e) {
+            throw new BusinessLogicException(MEMBER_NOT_EXIST);
+        }
+
+
+        Optional.ofNullable(patchDto.getUsername())
+                .ifPresent(foundMember::setUsername);
+        Optional.ofNullable(patchDto.getNewPassword())
+                .ifPresent(foundMember::setPassword);
+
+
+        return MemberResponseDto.of(memberRepository.save(foundMember));
+    }
 
 
     public Member findVerifiedMember(long memberId){
@@ -137,7 +171,14 @@ public class MemberService {
                 memberRepository.findById(memberId);
 
         return optionalMember.orElseThrow(() ->
-                new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
+                new BusinessLogicException(MEMBER_NOT_FOUND));
+    }
+
+
+    private void verifySameMember(Long preMemberId, Long editMemberId) {
+        if (!preMemberId.equals(editMemberId)) {
+            throw new BusinessLogicException(NOT_MATCH_TOKEN_WITH_MEMBER);
+        }
     }
 
     private RefreshToken verifyRefreshToken(String refreshTokenKey) {
@@ -145,5 +186,9 @@ public class MemberService {
                 .orElseThrow(() -> new BusinessLogicException(REFRESH_TOKEN_NOT_FOUND));
     }
 
+
+    public UsernamePasswordAuthenticationToken toAuthentication(String email, String password) {
+        return new UsernamePasswordAuthenticationToken(email, password);
+    }
 
 }
